Set up an Eclipse project for the robot.
New folder maple for Maple code.
New folder robot, which is an Eclipse Java Project.  The src
directory contains source code for the robot's main processing.
This is what we will run on the tablet.

jssc is added in robot/lib, and is now on the build path for the
project.  robot/src/RobotComm.java sets up a serial connection
but does not do anything else yet.  Right now, the robot just
prints "Hello world".

Play around with the project!